Update workedex.md
We're going to need to expand the length of the subset because all of the tweets were made before the funeral. Unless we want to look at some other date, it would be better to have the actual funeral tweets in there as well, but the dataset is huge so we still want a subset.