Update behind the scenes
I'm working on getting Docker to recognize unshrtn so I can run that without error. My system rebooted and uninstalled Docker so I have to set it up all over again. 
While I'm waiting for the command line to process, I'm going to update the python code for reading random lines from the file to create a subset of the dataset